이동 애니메이션 구현
1. 이동 시 해당 애니메이션 실행,
2. IDLE 상태일 경우 걷는모션 -> 가만히 서 있는 모션 변경,

해결할 사항 : 이미지 로드 문제로 다른 비트맵 인식 X, 제대로 로드하도록 변경

+++ zelda-2d/ImageManager.cpp
@@ -1,4 +1,5 @@
 ﻿#include "pch.h"
+
 #include "ImageManager.h"
 
 extern HINSTANCE hInst;
@@ -160,7 +161,7 @@ const string ImageManager::GetStringData(const BitmapKind kind, const int select
 
 AnimationObject* ImageManager::GetAnimationData(const int uiName)
 {
-	if (0 < animationData.size() && uiName < animationData.size())
+	if (0 < static_cast<int>(animationData.size()) && uiName < static_cast<int>(animationData.size()))
 		return &animationData[uiName];
 	else
 		return nullptr;
